fetcher fn simplified

+++ pages/api/activities/delete/[id].ts
@@ -7,9 +7,23 @@ export default validateRoute(async (req, res, user) => {
     res.json({ error: 'Email not verified' });
     return;
   }
+  // If method is not delete, throw error
+  if (req.method !== 'DELETE') {
+    res.status(405);
+    res.json({ error: 'Method not allowed' });
+    return;
+  }
+
+  const id = Number(req.query.id);
+  if (Number.isNaN(id)) {
+    res.status(400);
+    res.json({ error: 'Bad request' });
+    return;
+  }
+
   try {
     const deletedActivity = await prisma.activity.delete({
-      where: { id: req.body.id },
+      where: { id },
     });
     return res.status(200).json(deletedActivity);
   } catch (e) {
